+++ package.json
@@ -36,7 +36,6 @@
     "next": "13.1.2",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-query": "3.39.2",
     "type-check": "0.4.0",
     "typescript": "4.9.4"
   },
