+++ screen/survey/module/index.ts
@@ -1,0 +1,1 @@
+export * from './survey-content';
